requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
sqlalchemy>=1.4.0
pymysql>=1.0.2
cryptography>=3.4.7
python-dotenv>=0.19.0
pydantic>=1.8.0
